+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 0.2.0
 ### Breaking Changes
 * Updated protobuf file, `GetResponse.value` field type and `UpsertArgs.value` field type were changed from string to uint32
   * Users should use unit32 field type when are setting these values